A1200: First version. 38P_CLEAN

+++ include/asm-mips/sgiarcs.h
@@ -13,7 +13,6 @@
 #define _ASM_SGIARCS_H
 
 #include <linux/config.h>
-#include <asm/types.h>
 #include <asm/arc/types.h>
 
 /* Various ARCS error codes. */
@@ -116,7 +115,7 @@ enum arc_memtypes {
 	arc_prog,    /* A loaded program resides here */
 	arc_atmp,    /* temporary storage area */
 	arc_aperm,   /* permanent storage */
-	arc_fcontig, /* Contiguous and free */
+	arc_fcontig, /* Contiguous and free */    
 };
 
 union linux_memtypes {
@@ -165,11 +164,11 @@ enum linux_mountops {
 /* This prom has a bolixed design. */
 struct linux_bigint {
 #ifdef __MIPSEL__
-	u32 lo;
-	s32 hi;
+	unsigned long lo;
+	long hi;
 #else /* !(__MIPSEL__) */
-	s32 hi;
-	u32 lo;
+	long hi;
+	unsigned long lo;
 #endif
 };
